+++ wpilibc/src/main/native/cpp/simulation/XboxControllerSim.cpp
@@ -21,65 +21,65 @@ XboxControllerSim::XboxControllerSim(int port) : GenericHIDSim{port} {
 }
 
 void XboxControllerSim::SetLeftX(double value) {
-  SetRawAxis(static_cast<int>(XboxController::Axis::kLeftX), value);
+  SetRawAxis(XboxController::Axis::kLeftX, value);
 }
 
 void XboxControllerSim::SetRightX(double value) {
-  SetRawAxis(static_cast<int>(XboxController::Axis::kRightX), value);
+  SetRawAxis(XboxController::Axis::kRightX, value);
 }
 
 void XboxControllerSim::SetLeftY(double value) {
-  SetRawAxis(static_cast<int>(XboxController::Axis::kLeftY), value);
+  SetRawAxis(XboxController::Axis::kLeftY, value);
 }
 
 void XboxControllerSim::SetRightY(double value) {
-  SetRawAxis(static_cast<int>(XboxController::Axis::kRightY), value);
+  SetRawAxis(XboxController::Axis::kRightY, value);
 }
 
 void XboxControllerSim::SetLeftTriggerAxis(double value) {
-  SetRawAxis(static_cast<int>(XboxController::Axis::kLeftTrigger), value);
+  SetRawAxis(XboxController::Axis::kLeftTrigger, value);
 }
 
 void XboxControllerSim::SetRightTriggerAxis(double value) {
-  SetRawAxis(static_cast<int>(XboxController::Axis::kRightTrigger), value);
+  SetRawAxis(XboxController::Axis::kRightTrigger, value);
 }
 
 void XboxControllerSim::SetLeftBumper(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kBumperLeft), state);
+  SetRawButton(XboxController::Button::kBumperLeft, state);
 }
 
 void XboxControllerSim::SetRightBumper(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kBumperRight), state);
+  SetRawButton(XboxController::Button::kBumperRight, state);
 }
 
 void XboxControllerSim::SetLeftStickButton(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kStickLeft), state);
+  SetRawButton(XboxController::Button::kStickLeft, state);
 }
 
 void XboxControllerSim::SetRightStickButton(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kStickRight), state);
+  SetRawButton(XboxController::Button::kStickRight, state);
 }
 
 void XboxControllerSim::SetAButton(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kA), state);
+  SetRawButton(XboxController::Button::kA, state);
 }
 
 void XboxControllerSim::SetBButton(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kB), state);
+  SetRawButton(XboxController::Button::kB, state);
 }
 
 void XboxControllerSim::SetXButton(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kX), state);
+  SetRawButton(XboxController::Button::kX, state);
 }
 
 void XboxControllerSim::SetYButton(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kY), state);
+  SetRawButton(XboxController::Button::kY, state);
 }
 
 void XboxControllerSim::SetBackButton(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kBack), state);
+  SetRawButton(XboxController::Button::kBack, state);
 }
 
 void XboxControllerSim::SetStartButton(bool state) {
-  SetRawButton(static_cast<int>(XboxController::Button::kStart), state);
+  SetRawButton(XboxController::Button::kStart, state);
 }
